Reorganized client project
