+++ client/package.json
@@ -4,11 +4,12 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "axios": "^0.19.1",
     "bootstrap": "^4.3.1",
     "react": "^16.10.1",
     "react-dom": "^16.10.1",
     "react-router-dom": "^5.1.2",
-    "react-scripts": "3.1.2",
+    "react-scripts": "^3.3.0",
     "react-transition-group": "^4.3.0",
     "reactstrap": "^8.0.1",
     "uuid": "^3.3.3"
