+++ client/src/App.js
@@ -9,10 +9,10 @@ import {
   Route,
   Link
 } from "react-router-dom";
-import GiftCards from './components/GiftCards';
+import GiftCards from './screens/GiftCards';
 import Footer from './components/Footer';
-import GiftCard from './components/GiftCard';
-import CheckOut from './components/CheckoutConfirmation';
+import GiftCard from './screens/GiftCard';
+import CheckOut from './screens/CheckoutConfirmation';
 
 
 function App() {
